updates for cfel loader, not breaking tests
