+++ src/sed/loader/flash/buffer_handler.py
@@ -8,9 +8,12 @@ import dask.dataframe as dd
 import pyarrow.parquet as pq
 from joblib import delayed
 from joblib import Parallel
+from pandas import MultiIndex
 
 from sed.core.dfops import forward_fill_lazy
 from sed.core.logging import setup_logging
+from sed.loader.flash.dataframe import BaseDataFrameCreator
+from sed.loader.flash.dataframe import CFELDataFrameCreator
 from sed.loader.flash.dataframe import DataFrameCreator
 from sed.loader.flash.utils import get_channels
 from sed.loader.flash.utils import get_dtypes
@@ -40,11 +43,9 @@ class BufferFilePaths:
 
     def __init__(
         self,
-        config: dict,
         h5_paths: list[Path],
         folder: Path,
         suffix: str,
-        remove_invalid_files: bool,
     ) -> None:
         """Initializes the BufferFilePaths.
 
@@ -56,9 +57,6 @@ class BufferFilePaths:
         suffix = f"_{suffix}" if suffix else ""
         folder = folder / "buffer"
         folder.mkdir(parents=True, exist_ok=True)
-
-        if remove_invalid_files:
-            h5_paths = self.remove_invalid_files(config, h5_paths)
 
         self._file_paths = self._create_file_paths(h5_paths, folder, suffix)
 
@@ -93,18 +91,6 @@ class BufferFilePaths:
             return self._file_paths
         return [file_set for file_set in self if any(not file_set[key].exists() for key in DF_TYP)]
 
-    def remove_invalid_files(self, config, h5_paths: list[Path]) -> list[Path]:
-        valid_h5_paths = []
-        for h5_path in h5_paths:
-            try:
-                dfc = DataFrameCreator(config_dataframe=config, h5_path=h5_path)
-                dfc.validate_channel_keys()
-                valid_h5_paths.append(h5_path)
-            except InvalidFileError as e:
-                logger.info(f"Skipping invalid file: {h5_path.stem}\n{e}")
-
-        return valid_h5_paths
-
 
 class BufferHandler:
     """
@@ -125,13 +111,35 @@ class BufferHandler:
         self.n_cores: int = config["core"].get("num_cores", os.cpu_count() - 1)
         self.fp: BufferFilePaths = None
         self.df: dict[str, dd.DataFrame] = {typ: None for typ in DF_TYP}
+        fill_formats = self._config.get("fill_formats")
         self.fill_channels: list[str] = get_channels(
             self._config,
-            ["per_pulse", "per_train"],
+            fill_formats,
             extend_aux=True,
         )
         self.metadata: dict = {}
         self.filter_timed_by_electron: bool = None
+
+        core_beamline = config["core"].get("beamline")
+        self.DataFrameCreator: type[BaseDataFrameCreator] = None
+        if core_beamline == "pg2":
+            self.DataFrameCreator = DataFrameCreator
+        elif core_beamline == "cfel":
+            self.DataFrameCreator = CFELDataFrameCreator
+        else:
+            raise ValueError(f"Unsupported core beamline: {core_beamline}")
+
+    def _validate_h5_files(self, config, h5_paths: list[Path]) -> list[Path]:
+        valid_h5_paths = []
+        for h5_path in h5_paths:
+            try:
+                dfc = self.DataFrameCreator(config_dataframe=config, h5_path=h5_path)
+                dfc.validate_channel_keys()
+                valid_h5_paths.append(h5_path)
+            except InvalidFileError as e:
+                logger.info(f"Skipping invalid file: {h5_path.stem}\n{e}")
+
+        return valid_h5_paths
 
     def _schema_check(self, files: list[Path], expected_schema_set: set) -> None:
         """
@@ -190,7 +198,7 @@ class BufferHandler:
         logger.debug(f"Processing file: {paths['raw'].stem}")
         start_time = time.time()
         # Create DataFrameCreator and get get dataframe
-        df = DataFrameCreator(config_dataframe=self._config, h5_path=paths["raw"]).df
+        df = self.DataFrameCreator(config_dataframe=self._config, h5_path=paths["raw"]).df
 
         # Forward fill non-electron channels
         logger.debug(f"Forward filling {len(self.fill_channels)} channels")
@@ -205,6 +213,12 @@ class BufferHandler:
 
         # Create and save timed dataframe
         df_timed = self._create_timed_dataframe(df)
+        # timed dataframe
+        if isinstance(df.index, MultiIndex):
+            # drop the electron channels and only take rows with the first electronId
+            df_timed = df[self.fill_channels].loc[:, :, 0]
+        else:
+            df_timed = df[self.fill_channels]
         dtypes = get_dtypes(self._config, df_timed.columns.values)
         timed_df = df_timed.astype(dtypes).reset_index()
         logger.debug(f"Saving timed buffer with shape: {timed_df.shape}")
@@ -311,8 +325,11 @@ class BufferHandler:
         Returns:
             Tuple[dd.DataFrame, dd.DataFrame]: The electron and timed dataframes.
         """
-        self.fp = BufferFilePaths(self._config, h5_paths, folder, suffix, remove_invalid_files)
         self.filter_timed_by_electron = filter_timed_by_electron
+        if remove_invalid_files:
+            h5_paths = self._validate_h5_files(self._config, h5_paths)
+
+        self.fp = BufferFilePaths(h5_paths, folder, suffix)
 
         if not force_recreate:
             schema_set = set(
